OOPify one method
