+++ src/main/java/ee/tuleva/onboarding/mandate/batch/MandateBatchService.java
@@ -1,15 +1,12 @@
 package ee.tuleva.onboarding.mandate.batch;
 
-import static ee.tuleva.onboarding.mandate.MandateType.PARTIAL_WITHDRAWAL;
 import static ee.tuleva.onboarding.mandate.response.MandateSignatureStatus.*;
 import static ee.tuleva.onboarding.mandate.response.MandateSignatureStatus.SIGNATURE;
 import static ee.tuleva.onboarding.pillar.Pillar.SECOND;
-import static ee.tuleva.onboarding.pillar.Pillar.THIRD;
 import static java.util.stream.Collectors.toList;
 
 import ee.tuleva.onboarding.auth.principal.AuthenticatedPerson;
 import ee.tuleva.onboarding.epis.EpisService;
-import ee.tuleva.onboarding.epis.mandate.details.PartialWithdrawalMandateDetails;
 import ee.tuleva.onboarding.error.response.ErrorResponse;
 import ee.tuleva.onboarding.error.response.ErrorsResponse;
 import ee.tuleva.onboarding.mandate.MandateFileService;
@@ -80,7 +77,7 @@ public class MandateBatchService {
           throw new IllegalArgumentException(
               "Can only create third pillar withdrawals before retirement age");
         }
-      } else if (!isBatchOnlyThirdPillarPartialWithdrawal(mandateBatchDto)
+      } else if (!mandateBatchDto.isBatchOnlyThirdPillarPartialWithdrawal()
           && !eligibility.hasReachedEarlyRetirementAge()) {
         throw new IllegalArgumentException(
             "Can only do partial withdrawal from III pillar before early retirement age");
@@ -134,24 +131,6 @@ public class MandateBatchService {
     }
 
     return getBatchProcessingStatusAndHandleIfProcessed(user, mandateBatch, locale);
-  }
-
-  private boolean isBatchOnlyThirdPillarPartialWithdrawal(MandateBatchDto mandateBatchDto) {
-    var mandates = mandateBatchDto.getMandates();
-
-    if (mandates.size() > 1) {
-      return false;
-    }
-
-    var mandateDto = mandates.getFirst();
-
-    if (mandateDto.getMandateType() != PARTIAL_WITHDRAWAL) {
-      return false;
-    }
-
-    var details = (PartialWithdrawalMandateDetails) mandateDto.getDetails();
-
-    return details.getPillar() == THIRD;
   }
 
   private MandateSignatureStatus persistFileSignedWithIdCard(
